pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kleinanzeigen-bot"
version = "0.1.0"
description = "KI-gestützter Bot zum automatischen Erstellen von Kleinanzeigen"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "ollama>=0.4.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "Pillow>=10.4.0",
    "playwright>=1.48.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "respx>=0.21.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["ollama.*", "bs4.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "e2e: marks end-to-end tests",
]

[tool.coverage.run]
source = ["src/kleinanzeigen_bot"]

[tool.coverage.report]
fail_under = 80

[tool.hatch.build.targets.wheel]
packages = ["src/kleinanzeigen_bot"]
